BuildImage(): check actual commit against develop
In BuildImage(), we check whether build_oai, tools/build_helper, or
Dockerfile.base changed to decide whether we need to rebuild the base
image (ran-base). We checked using HEAD, but already merged develop in.
Hence, this check might fail, and in this commit, we diff the commit ID
to compile agains develop.